Start refactoring our event dispatching/handling plumbing to cut boilerplate.
Make the event handlers responsible for both decoding the event and
deciding if the event should be handled by the handler. The dispatch
code is becoming trivial and a candidate for more refactoring.

The event handlers standardize on a calling pattern taking the current
redux state (the whole state) and the event in question.

Event handlers always return a promise.

We have some extracted common code from the event handlers. The whole
"returning a no-op resolution when the event is ignored" is a smell to
be fixed later. Strongly suspect we'll be able to invert the dependency
and make the event handlers not know about the common code as the
design emerges.

The signal dependency looks to be a candidate for removal as we don't
have a need for adding and removing arbitrary listeners all over the
place.