Add first react view test for hello world view.
Start using enzyme (https://github.com/airbnb/enzyme) for react view
tests. We use shallow rendering to test one and only one component
and not any component/view it may contain.

Split the hello world (incrementing counter) view from its react-redux
wiring as it made understanding the enzyme-driven tests for the simple
component difficult. We test the real view and leave testing the
react-redux configuration/wiring for later.

What do we think our test does? It verifies that when you click on its
button, a particular event is fired on the application event bus. It
also checks to see if a value given in as a particular property gets
rendered somewhere in the element.

Initial step toward useful view code test coverage for #3.